run node-gyp in the postinstall phase
if `preinstall` script is not set, then `npm` will default
that to `node-gyp rebuild` *if* there is a binding.gyp
file present.

This is problematic because:

If we want to run node-gyp on as many cores as possible,
we can do that by simply overriding the `preinstall` script,
like:

```json
  "scripts": {
    "preinstall": "JOBS=max node-gyp rebuild"
  }
```

But since we depend on `nan` to exist for doing this (we
build up the include path by require('nan')) this fails,
since nan is not there if the node_modules folder is empty.

By moving the build to `postinstall` we can know for sure
that `nan` will be there before we run `node-gyp`. To prevent
`npm` from not running default version of `preinstall`, we just
run `true` (yes a bit hacky)